+++ api/src/main/java/com/wissem/user/UserRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
   Optional<User> findByEmail(String email);
+
   boolean existsByEmail(String email);
 }
